refactor: event.dto 패키지 분리
