+++ core/core-domain/src/main/java/com/woopaca/taximate/core/domain/event/ParticipationEventProducer.java
@@ -1,5 +1,7 @@
 package com.woopaca.taximate.core.domain.event;
 
+import com.woopaca.taximate.core.domain.event.dto.LeaveEvent;
+import com.woopaca.taximate.core.domain.event.dto.ParticipateEvent;
 import com.woopaca.taximate.core.domain.party.Party;
 import com.woopaca.taximate.core.domain.user.User;
 import org.springframework.context.ApplicationEventPublisher;
