+++ core/core-domain/src/main/java/com/woopaca/taximate/core/domain/event/dto/LeaveEvent.java
@@ -1,4 +1,4 @@
-package com.woopaca.taximate.core.domain.event;
+package com.woopaca.taximate.core.domain.event.dto;
 
 import com.woopaca.taximate.core.domain.party.Party;
 import com.woopaca.taximate.core.domain.user.User;
